Implemented ResultBox
Added RollType enum to src/dice.vala
Added static member instances of Die corresponding to the dice most commonly used in tabletop.
Updated doc comments with the @link tag (hopefully did so correctly).

Moved DiceButtons.Spacing constant over to App.DefaultSpacing so it can be a bit more generic.

Created ResultBox widget to contain the UI elements for displaying the current equation as well as the output after the roll is executed.
Added src/ui/result.vala to meson.build.

Removed the close button from MainWindow; It's superfluous and not going to be in the final application.
Reorganized the widgets in MainWindow to make room to have ResultBox positioned above the button row.
Added roll, takeHighest, and takeLowest buttons which will execute the roll with their corresponding RollType.
Wired up the three rolling buttons to test methods to test and validate the functionality of ResultBox.