Iterated on ResultBox
Removed the results TreeMap, as it's entirely unnecessary. Instead, since we have the Die and the quantity, simply do the rolls on-demand when building the output string. This will prevent replicating results but that's not really what this application is supposed to do, so it's not a real loss of functionality.